requirements.txt
streamlit
langchain-core
langchain-groq
langchain-chroma==0.1.4
langchain-huggingface
langchain-ollama
langchain-text-splitters
PyPDF
chromadb